Fixed problem with using worldline in classic mode on Linux
When using worldline in classic mode, the resampler path is the
path to the worldline.so file, but with the code from my previous
patch that got external wavtools working, that file path was being
passed into ResolveResamplerExePathLinux as if it were a shell
wrapper and not an ELF executable. The mathod found something in
the file that matched the criteria and pulled that out as the "path
to the exe file".

This new fix adds a check to ResolveResamplerExePathLinux to see if
the file is an ELF file, and if so, return the path to the ELF
file. It was odd to me at first that this works, but I tested it
outside of OpenUtau, and it turns out that a batch file running
through wine on linux can run ELF executables so long as they have
a .so file extension.